Added Categories
Added categories and add record and delete record to the data facade

+++ finance_tracker/src/main/java/com/finance_tracker/account/AccountRecord.java
@@ -1,10 +1,10 @@
 package com.finance_tracker.account;
 
-import com.finance_tracker.Date;
+import java.util.Date;
 
 public class AccountRecord {
     private Date date;
-    private double amount;
+    private Double amount;
 
 
     /**
@@ -26,7 +26,7 @@ public class AccountRecord {
     /**
      * @return float
      */
-    public double getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
@@ -34,7 +34,7 @@ public class AccountRecord {
     /**
      * @param amount float
      */
-    public void setAmount(float amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
 }
